Fix. 헤더 구조 변경 - gnb제거 - 검색버튼과 프로필 이미지만 남겨둠
