--- src/constants/headerMenus.js
@@ -1,26 +1,0 @@
-// import cart from "../images/ico-cart.svg";
-import search from "../images/ico-search.svg";
-export const menus = [
-  {
-    title: "홈",
-    path: "/",
-    iconPath: null,
-  },
-  // {
-  //   title: "예약",
-  //   path: null,
-  //   iconPath: null,
-  //   sectionId: "reservation",
-  // },
-  // {
-  //   title: null,
-  //   path: "my/cart",
-  //   iconPath: cart,
-  // },
-  {
-    title: null,
-    path: null,
-    iconPath: search,
-    sectionId: "search",
-  },
-];
