+++ src/scss/layout/_header.scss
@@ -22,6 +22,7 @@
 
 .header__inner {
   display: flex;
+  gap: 1.6rem;
   align-items: center;
 }
